requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
PyQt5==5.13.0
PyQt5-sip==4.19.18
requests==2.22.0
urllib3==1.25.6
